backend-flask/requirements.txt
flask
psycopg2-binary
chromadb
sentence-transformers
requests
pyjwt